requirements.txt
streamlit
snowflake-connector-python
sseclient-py
requests
